+++ ItlizeSharpJoole/Joole.Repo/ISubCategoryRepo.cs
@@ -1,0 +1,16 @@
+﻿using Joole.DAL;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Joole.Repo
+{
+    public interface ISubCategoryRepo
+    {
+        IEnumerable<SubCategory> GetSubCategories();
+        SubCategory GetSubCategory(int id);
+        IEnumerable<SubCategory> getSubforCategory(int id);
+    }
+}
