+++ ItlizeSharpJoole/Joole.Services/Joole.Services.csproj
@@ -93,9 +93,9 @@
   <ItemGroup>
     <Compile Include="App_Start\UnityConfig.cs" />
     <Compile Include="App_Start\UnityMvcActivator.cs" />
-    <Compile Include="Class1.cs" />
     <Compile Include="FilterService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ResultsService.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
