+++ ItlizeSharpJoole/Joole.UI/App_Start/RouteConfig.cs
@@ -18,6 +18,18 @@ namespace Joole.UI
                 url: "{controller}/{action}/{id}",
                 defaults: new { controller = "Home", action = "Index", id = UrlParameter.Optional }
             );
+
+            routes.MapRoute(
+                name: "Results",
+                url: "{controller}/{action}/{id}",
+                defaults: new { controller = "Results", action = "Results", id = UrlParameter.Optional }
+            );
+
+            routes.MapRoute(
+                name: "SearchResults",
+                url: "{controller}/{action}/{id}",
+                defaults: new { controller = "Results", action = "SearchResults", id = UrlParameter.Optional }
+            );
         }
     }
 }
