+++ ItlizeSharpJoole/Joole.UI/Content/Site.css
@@ -1,5 +1,4 @@
 ﻿body {
-    padding-top: 50px;
     padding-bottom: 20px;
 }
