+++ ItlizeSharpJoole/Joole.UI/Scripts/jquery-3.4.1.intellisense.js
