+++ ItlizeSharpJoole/Joole.UI/Views/Home/Index.cshtml
@@ -12,7 +12,9 @@
     <div class="col-md-4">
         @ViewBag.category <br />
         @ViewBag.categories[3].CategoryName
+        <h1>Test</h1>
         <h2>Getting started</h2>
+
         <p>
             ASP.NET MVC gives you a powerful, patterns-based way to build dynamic websites that
             enables a clean separation of concerns and gives you full control over markup
