+++ ItlizeSharpJoole/Joole.UI/Views/Results/SearchResults.cshtml
@@ -1,0 +1,49 @@
+﻿@model Joole.UI.Models.ResultsVM
+@{
+    ViewBag.Title = "SearchResults";
+}
+
+<div style="background-image: linear-gradient(#7e7ed0, white)" class="row">
+    @{ Html.RenderPartial("_SearchBar"); }
+</div>
+<div class="row">
+
+    <div style="border: 2px solid black; padding: 10px;" class="col-lg-3">
+        <div class="row">
+            <div class="col-lg-4"><p>Search:</p></div>
+            <div class="col-lg-4"><button class="btn-primary">Save</button></div>
+            <div class="col-lg-4"><button class="btn-primary">Clear</button></div>
+        </div>
+        <div>
+
+        </div>
+        <br />
+        <div class="row">
+            <div style="padding-right: 0px;" class="col-lg-3"><p>Model Year</p></div>
+            <div class="col-lg-9 text-center">
+                <div class="col-lg-4"><input style="width:60px;" type="number" /></div>
+                <div class="col-lg-1"><p>-</p></div>
+                <div class="col-lg-4"><input style="width:60px;" type="number" /></div>
+            </div>
+        </div><br />
+        <div>
+            <button class="btn btn-primary" type="button">
+                Dropdown Example
+                <span class="caret"></span>
+            </button>
+        </div>
+        <div id="techSpecs">
+            @foreach(var item in Model.SpecList)
+            {
+                <p>@item</p>
+            }
+        </div>
+    </div>
+
+
+    <div style="border:2px solid green" class="col-lg-9">
+
+    </div>
+</div>
+
+
