+++ ItlizeSharpJoole/Joole.UI/Views/Shared/_SearchBar.cshtml
@@ -1,0 +1,54 @@
+﻿<div style="height: 120px; background-image: linear-gradient(#7e7ed0, white)" class="container-fluid text-center" id="searchBar">
+    <p style="padding-top: 25px; font-size: 45px; display:inline-block; float: left;">jOOle</p>
+    <div style="padding-top: 45px;">
+
+        @Html.DropDownList("selectedCategory", new SelectList(ViewBag.categories), "Categories" , new { @id="selectedCategory", @onchange="categoryChanged()", @style="height: 26px" })
+        <input id="subCategoryText" list="subs" style="width: 60%;" type="text" />
+        <datalist id="subs"></datalist>
+        <button style="background-color: #4949e6; color: white; padding: 5px;" class="glyphicon glyphicon-search"onclick="GetInput()"></button>
+    </div>
+</div>
+
+<script>
+    function categoryChanged() {
+        $('#selectedCategory').change(function () {
+            var selectedCategory = $('#selectedCategory').val();
+            //alert(selectedCategory);
+            if (selectedCategory != "") {
+                var selectedSubCategory = $('#selectedSubCategory');
+                selectedSubCategory.empty();
+                $('#subCategoryText').val('');
+                $.getJSON('/Results/GetSubCategories', { category: selectedCategory }, function (subcategories) {
+                    $('#subs').empty();
+                    $.each(subcategories, function (index, sub) {
+                        $('#subs').append($('<option/>', {
+                            value: sub.Text,
+                            id: sub.Value
+                        }));
+                    });
+                });
+            } else {
+                $('#subs').empty();
+            }
+        });
+    }
+        
+
+    function GetInput() {
+        var category = $('#selectedCategory').val();
+        var subCategory = $('#subCategoryText').val();
+        //alert($("#subs option[value='" + $('#subCategoryText').val() + "']").attr('id'));
+        var id = $("#subs option[value='" + $('#subCategoryText').val() + "']").attr('id');
+        //alert(id);
+        $.ajax({
+            url: '/Results/SearchResults/',
+            type: "POST",
+            dataType: "JSON",
+            data: { Category: category, SubCategory: subCategory, SubCategoryID: id.toString() },
+            success: function (data) {
+                window.location.href = data.url;
+            }
+        })
+    }
+
+</script>
