+++ ItlizeSharpJoole/Joole.UI/packages.config
@@ -3,7 +3,7 @@
   <package id="Antlr" version="3.5.0.2" targetFramework="net472" />
   <package id="bootstrap" version="3.4.1" targetFramework="net472" />
   <package id="EntityFramework" version="6.3.0" targetFramework="net472" />
-  <package id="jQuery" version="3.3.1" targetFramework="net472" />
+  <package id="jQuery" version="3.4.1" targetFramework="net472" />
   <package id="jQuery.Validation" version="1.17.0" targetFramework="net472" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net472" />
   <package id="Microsoft.AspNet.Razor" version="3.2.7" targetFramework="net472" />
